[*] Add accepting default values
When generating a new ca the script would ask the user for a bunch of
information on the distinguished name. If you wanted to suppress asking
for defaults you had to specify all the values `-c`, `-st`, `-l`, `-o`
and `-cn` even if you wanted to use the defaults.

This commit adds the `--accept` flag. When passed the script will use
the default value when non was provided on the commandline  without
asking for user input.